Permission projek RnD

+++ resources/views/livewire/projek-rnd-table.blade.php
@@ -42,11 +42,11 @@
                     @include('livewire.dataperpage')
                 </li>
                 <li class="m-1">
-                    {{-- @can('tambah-projek') --}}
+                    @can('tambah-projek-rnd')
                         <a href="{{ route('projek-rnd.create') }}" class="mt-2 block w-fit rounded-md py-1.5 px-2 bg-indigo-600 text-sm font-semibold text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600">
                             Tambah
                         </a>
-                    {{-- @endcan --}}
+                    @endcan
                 </li>
             </ul>
         </div>
@@ -79,9 +79,9 @@
                             </td>
                             <td class="px-6 py-3">
                                 @if ($projek_rnd->status !== 'Konfirmasi')
-                                {{-- @can('edit-projek') --}}
+                                @can('edit-projek-rnd')
                                     <strong><u><a href="{{ route('projek-rnd.edit', $projek_rnd->id) }}">{{ $projek_rnd->kode_projek_rnd }}</a></u></strong>
-                                {{-- @endcan --}}
+                                @endcan
                                 @else
                                     <span class="text-gray-500">{{ $projek_rnd->kode_projek_rnd }}</span>
                                 @endif
@@ -95,18 +95,20 @@
                             <td class="px-6 py-4">
                                 <div class="row flex space-x-2">
                                     @if ($projek_rnd->status === 'Konfirmasi')
-                                        <button wire:click="deleteProjekRnd({{ $projek_rnd->id }})"
-                                            data-modal-target="deleteprojekrnd-modal" data-modal-toggle="deleteprojekrnd-modal"
-                                            class="rounded-md border border-slate-300 py-1 px-2 text-center text-xs transition-all shadow-sm hover:shadow-lg text-slate-600 hover:text-white hover:bg-red-600 hover:border-red-600 focus:text-white focus:bg-red-600 focus:border-red-600 active:border-red-600 active:text-white active:bg-red-600 disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none"
-                                            type="button">
-                                            <svg class="w-[16px] h-[16px] text-gray-800 dark:text-white" aria-hidden="true"
-                                                xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none"
-                                                viewBox="0 0 24 24">
-                                                <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round"
-                                                    stroke-width="2"
-                                                    d="M5 7h14m-9 3v8m4-8v8M10 3h4a1 1 0 0 1 1 1v3H9V4a1 1 0 0 1 1-1ZM6 7h12v13a1 1 0 0 1-1 1H7a1 1 0 0 1-1-1V7Z" />
-                                            </svg>
-                                        </button>
+                                        @can('hapus-projek-rnd')
+                                            <button wire:click="deleteProjekRnd({{ $projek_rnd->id }})"
+                                                data-modal-target="deleteprojekrnd-modal" data-modal-toggle="deleteprojekrnd-modal"
+                                                class="rounded-md border border-slate-300 py-1 px-2 text-center text-xs transition-all shadow-sm hover:shadow-lg text-slate-600 hover:text-white hover:bg-red-600 hover:border-red-600 focus:text-white focus:bg-red-600 focus:border-red-600 active:border-red-600 active:text-white active:bg-red-600 disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none"
+                                                type="button">
+                                                <svg class="w-[16px] h-[16px] text-gray-800 dark:text-white" aria-hidden="true"
+                                                    xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none"
+                                                    viewBox="0 0 24 24">
+                                                    <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round"
+                                                        stroke-width="2"
+                                                        d="M5 7h14m-9 3v8m4-8v8M10 3h4a1 1 0 0 1 1 1v3H9V4a1 1 0 0 1 1-1ZM6 7h12v13a1 1 0 0 1-1 1H7a1 1 0 0 1-1-1V7Z" />
+                                                </svg>
+                                            </button>
+                                        @endcan
                                     @endif
                                 </div>
                             </td>
